+++ packages/eyes-selenium/scripts/render.js
@@ -137,6 +137,7 @@ if (!url) {
     configuration.setProxy(args.proxy)
   }
   eyes.setConfiguration(configuration)
+
   const {logger, logFilePath} = initLog(eyes, new URL(url).hostname.replace(/\./g, '-'))
   logger.log('[render script] Running Selenium render for', url)
   logger.log(`[render script] process versions: ${JSON.stringify(process.versions)}`)